feat: rename singUp to accessForm
